fix:fix the handling for escape.
