+++ cmake/config.cmake
@@ -1,4 +1,4 @@
 set(CMAKE_BUILD_TYPE RelWithDebInfo)
 set(XGRAMMAR_BUILD_PYTHON_BINDINGS ON)
-set(XGRAMMAR_BUILD_CXX_TESTS OFF)
+set(XGRAMMAR_BUILD_CXX_TESTS ON)
 set(XGRAMMAR_ENABLE_CPPTRACE OFF)
